refactor buffer loader
