+++ crates/cubecl-linalg/src/matmul/components/global/loader/sync/buffer/mod.rs
@@ -1,0 +1,3 @@
+mod cyclic;
+
+pub use cyclic::*;
