+++ crates/cubecl-linalg/src/matmul/components/global/loader/sync/full/mod.rs
@@ -1,0 +1,7 @@
+mod cyclic;
+mod strided;
+mod tilewise;
+
+pub use cyclic::*;
+pub use strided::*;
+pub use tilewise::*;
