+++ crates/cubecl-linalg/src/matmul/components/global/multi_stage/specialized/loader.rs
@@ -1,7 +1,7 @@
 use std::marker::PhantomData;
 
 use crate::matmul::components::global::base::GlobalConfig as _;
-use crate::matmul::components::global::loader::sync::SyncLoadingStrategy;
+use crate::matmul::components::global::loader::sync::SyncBufferLoadingStrategy;
 use crate::matmul::components::global::tensor_view::TensorReader;
 use crate::matmul::components::global::{InputLoader, SyncInputLoader};
 use crate::matmul::components::stage::single_buffer::{LhsBufferReader, RhsBufferReader};
@@ -18,7 +18,7 @@ pub struct SyncLhsBufferLoader<
     EG: Numeric,
     ES: Numeric,
     S: stage::StageConfig,
-    L: SyncLoadingStrategy,
+    L: SyncBufferLoadingStrategy,
 > {
     pub tensor_view: TensorReader<EG>,
     pub stage: Stage<ES, L::TilingLayout>,
@@ -33,7 +33,7 @@ pub struct SyncRhsBufferLoader<
     EG: Numeric,
     ES: Numeric,
     S: stage::StageConfig,
-    L: SyncLoadingStrategy,
+    L: SyncBufferLoadingStrategy,
 > {
     pub tensor_view: TensorReader<EG>,
     pub stage: Stage<ES, L::TilingLayout>,
@@ -44,7 +44,7 @@ pub struct SyncRhsBufferLoader<
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncLoadingStrategy>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
     InputLoader<EG, ES, Config<S>> for SyncLhsBufferLoader<EG, ES, S, L>
 {
     type StageReader = LhsBufferReader<ES, L::TilingLayout>;
@@ -64,7 +64,7 @@ impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncLoadingStrategy>
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncLoadingStrategy>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
     SyncInputLoader<EG, ES, Config<S>> for SyncLhsBufferLoader<EG, ES, S, L>
 {
     fn fill_stage(this: &mut Self, #[comptime] config: Config<S>) {
@@ -81,7 +81,7 @@ impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncLoadingStrategy>
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncLoadingStrategy>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
     SyncLhsBufferLoader<EG, ES, S, L>
 {
     pub fn new(
@@ -107,7 +107,7 @@ impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncLoadingStrategy>
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncLoadingStrategy>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
     InputLoader<EG, ES, Config<S>> for SyncRhsBufferLoader<EG, ES, S, L>
 {
     type StageReader = RhsBufferReader<ES, L::TilingLayout>;
@@ -127,7 +127,7 @@ impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncLoadingStrategy>
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncLoadingStrategy>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
     SyncInputLoader<EG, ES, Config<S>> for SyncRhsBufferLoader<EG, ES, S, L>
 {
     fn fill_stage(this: &mut Self, #[comptime] config: Config<S>) {
@@ -144,7 +144,7 @@ impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncLoadingStrategy>
 }
 
 #[cube]
-impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncLoadingStrategy>
+impl<EG: Numeric, ES: Numeric, S: stage::StageConfig, L: SyncBufferLoadingStrategy>
     SyncRhsBufferLoader<EG, ES, S, L>
 {
     pub fn new(
